Move PlutusData under versioned modules
