+++ plutus-ledger-api/src/lib.rs
@@ -2,7 +2,6 @@ pub(crate) mod feature_traits;
 pub mod generators;
 #[cfg(feature = "lbf")]
 pub mod lamval;
-pub mod plutus_data;
 pub mod v1;
 pub mod v2;
 #[cfg(feature = "lbf")]
