+++ plutus-ledger-api/src/v1/address.rs
@@ -1,8 +1,8 @@
 //! Types related to Cardano addresses
-use crate::plutus_data::{
+use crate::v1::crypto::Ed25519PubKeyHash;
+use crate::v1::plutus_data::{
     verify_constr_fields, IsPlutusData, PlutusData, PlutusDataError, PlutusType,
 };
-use crate::v1::crypto::Ed25519PubKeyHash;
 use crate::v1::script::ValidatorHash;
 #[cfg(feature = "lbf")]
 use lbr_prelude::json::{self, Error, Json};
