+++ plutus-ledger-api/src/v2/mod.rs
@@ -9,6 +9,7 @@ pub use crate::v1::address;
 pub use crate::v1::assoc_map;
 pub use crate::v1::crypto;
 pub use crate::v1::interval;
+pub use crate::v1::plutus_data;
 pub use crate::v1::redeemer;
 pub use crate::v1::script;
 pub use crate::v1::value;
