+++ plutus-ledger-api/src/v2/transaction.rs
@@ -1,12 +1,13 @@
 //! Types related to Cardano transactions.
-use crate::plutus_data::{parse_constr_with_tag, parse_fixed_len_constr_fields};
-use crate::plutus_data::{
-    verify_constr_fields, IsPlutusData, PlutusData, PlutusDataError, PlutusType,
-};
+
 #[cfg(feature = "chrono")]
 pub use crate::v1::transaction::POSIXTimeConversionError;
 pub use crate::v1::transaction::{
     DCert, POSIXTime, POSIXTimeRange, ScriptPurpose, TransactionHash, TransactionInput,
+};
+use crate::v2::plutus_data::{
+    parse_constr_with_tag, parse_fixed_len_constr_fields, verify_constr_fields, IsPlutusData,
+    PlutusData, PlutusDataError, PlutusType,
 };
 #[cfg(feature = "lbf")]
 use lbr_prelude::json::Json;
